Guess the correct word or die
It will randomly generate a word and you have to guess a letter and if the letter is in the word then you can choose another letter and if it is not in the word then one life will be taken away. You have 6 lives means 6 chance to guess the correct word.